pyproject.toml
[tool.poetry]
name = "code"
version = "0.1.0"
description = ""
authors = ["Takaaki Suzuki <test@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
mysqlclient = "^2.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
